+++ src/Instrumentation/Laravel/src/ConsoleInstrumentation.php
@@ -1,0 +1,90 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OpenTelemetry\Contrib\Instrumentation\Laravel;
+
+use Illuminate\Console\Command;
+use Illuminate\Contracts\Console\Kernel;
+use OpenTelemetry\API\Instrumentation\CachedInstrumentation;
+use OpenTelemetry\API\Trace\Span;
+use OpenTelemetry\API\Trace\SpanKind;
+use OpenTelemetry\API\Trace\StatusCode;
+use OpenTelemetry\Context\Context;
+use function OpenTelemetry\Instrumentation\hook;
+use OpenTelemetry\SemConv\TraceAttributes;
+use Throwable;
+
+class ConsoleInstrumentation
+{
+    public static function register(CachedInstrumentation $instrumentation): void
+    {
+        hook(
+            Kernel::class,
+            'handle',
+            pre: static function (Kernel $kernel, array $params, string $class, string $function, ?string $filename, ?int $lineno) use ($instrumentation) {
+                /** @psalm-suppress ArgumentTypeCoercion */
+                $builder = $instrumentation->tracer()
+                    ->spanBuilder('Artisan handler')
+                    ->setSpanKind(SpanKind::KIND_PRODUCER)
+                    ->setAttribute(TraceAttributes::CODE_FUNCTION, $function)
+                    ->setAttribute(TraceAttributes::CODE_NAMESPACE, $class)
+                    ->setAttribute(TraceAttributes::CODE_FILEPATH, $filename)
+                    ->setAttribute(TraceAttributes::CODE_LINENO, $lineno);
+                $parent = Context::getCurrent();
+                $span = $builder->startSpan();
+                Context::storage()->attach($span->storeInContext($parent));
+
+                return $params;
+            },
+            post: static function (Kernel $kernel, array $params, ?int $exitCode, ?Throwable $exception) {
+                $scope = Context::storage()->scope();
+                if (!$scope) {
+                    return;
+                }
+                $scope->detach();
+                $span = Span::fromContext($scope->context());
+                if ($exception) {
+                    $span->recordException($exception, [TraceAttributes::EXCEPTION_ESCAPED => true]);
+                    $span->setStatus(StatusCode::STATUS_ERROR, $exception->getMessage());
+                } elseif ($exitCode !== Command::SUCCESS) {
+                    $span->setStatus(StatusCode::STATUS_ERROR);
+                } else {
+                    $span->setStatus(StatusCode::STATUS_OK);
+                }
+
+                $span->end();
+            }
+        );
+
+        hook(
+            Command::class,
+            'execute',
+            pre: static function (Command $command, array $params, string $class, string $function, ?string $filename, ?int $lineno) use ($instrumentation) {
+                $scope = Context::storage()->scope();
+                $span = Span::fromContext($scope->context());
+                $span->addEvent('command starting', [
+                    'command' => $command->getName(),
+                ]);
+
+                return $params;
+            },
+            post: static function (Command $command, array $params, ?int $exitCode, ?Throwable $exception) {
+                $scope = Context::storage()->scope();
+                if (!$scope) {
+                    return;
+                }
+                $span = Span::fromContext($scope->context());
+                $span->addEvent('command finished', [
+                    'command' => $command->getName(),
+                    'exit-code' => $exitCode,
+                ]);
+
+                if ($exception) {
+                    $span->recordException($exception, [TraceAttributes::EXCEPTION_ESCAPED => true]);
+                    $span->setStatus(StatusCode::STATUS_ERROR, $exception->getMessage());
+                }
+            }
+        );
+    }
+}
